backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Async Task Queue
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
flower==2.0.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0
feedparser==6.0.10

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
structlog==24.1.0
sentry-sdk==1.39.1
psutil==5.9.7

# API Documentation
drf-spectacular==0.27.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
